requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
Pillow==9.2.0
py==1.11.0
python-dotenv==0.21.0
python-i18n==0.3.9
pytz==2022.5
PyYAML==6.0
yarl==1.8.1